Get patches from pool for pipeline
Create method get_patches_pipeline() for PatchPool that gets a list of
the indices of a certain image (identified by tfRecord path).
Depending on the status of the patches and the mode of PatchPool (build
pipeline for querying or training with AL) the needed patch indices are
chosen from the entire pool.
Implement the function in pipeline.py