Save data converted to hdf5 in useful structure
Change the conversion function that converts tfRecord and saves it as
hdf5 file, in a way so that the hdf5 file data structure is useful for
later data reading. The data is now organized in image and label data,
the image data is further divided into training and validation data. The
 ids of the images and accompanying labels are the same